third_party/blink/web_tests/media/video-play-pause-exception.html

<!DOCTYPE html>
<title>Test that the playing event is not fired when video has no src.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.onplaying = t.unreached_func();
    video.onloadstart = t.step_func(function() {});
    video.ontimeupdate = t.step_func(function() {});
    video.onwaiting = t.step_func(function() {});

    video.onpause = t.step_func_done(function() {
        assert_equals(video.networkState, HTMLMediaElement.NETWORK_EMPTY);
    });

    video.play();
    video.pause();
});
</script>
